+++ src/docker-code-server/main.tf
@@ -50,7 +50,7 @@ data "coder_parameter" "image" {
   type        = "string"
   description = "What container image and language do you want?"
   mutable     = true
-  default     = "marktmilligan/python-ai-agents:latest"
+  default     = "marktmilligan/go:latest"
   icon        = "https://www.docker.com/wp-content/uploads/2022/03/vertical-logo-monochromatic.png"
 
   option {
@@ -77,7 +77,7 @@ data "coder_parameter" "repo" {
   description = "What source code repository do you want to clone?"
   mutable     = true
   icon        = "/icon/git.svg"
-  default     = ""
+  default     = "https://github.com/coder/coder"
   option {
     name = "Do not clone a repository"
     value = ""
@@ -101,41 +101,33 @@ data "coder_parameter" "repo" {
   order       = 2
 }
 
-data "coder_parameter" "dotfiles_url" {
-  name        = "Dotfiles URL (optional)"
-  description = "Personalize your workspace e.g., https://github.com/sharkymark/dotfiles.git"
-  type        = "string"
-  default     = ""
-  mutable     = true
-  icon        = "https://git-scm.com/images/logos/downloads/Git-Icon-1788C.png"
-  order       = 3
+module "dotfiles" {
+  count    = data.coder_workspace.me.start_count
+  source   = "registry.coder.com/modules/dotfiles/coder"
+  agent_id = coder_agent.dev.id
 }
 
-data "coder_parameter" "ide" {
-  name        = "VS Code IDE"
-  description = "Select a local or browser-based IDE"
-  type        = "string"
-  default     = "code"
-  mutable     = true
-  icon        = "/icon/code.svg"
-  order       = 5
+module "vscode-web" {
+  count          = data.coder_workspace.me.start_count
+  source         = "registry.coder.com/modules/vscode-web/coder"
+  agent_id       = coder_agent.dev.id
+  extensions     = ["github.copilot"]
+  folder         = "/home/coder"
+  accept_license = true
+}
 
-  option {
-    name = "VS Code Desktop"
-    value = "code"
-    icon = "/icon/code.svg"
-  }
-  option {
-    name = "code-server (browser IDE)"
-    value = "code-server"
-    icon = "/icon/coder.svg"
-  }
-  option {
-    name = "Zed (desktop IDE)"
-    value = "zed"
-    icon = "/icon/zed.svg"
-  }
+module "git-clone" {
+  count    = data.coder_workspace.me.start_count
+  source   = "registry.coder.com/modules/git-clone/coder"
+  version  = "1.0.18"
+  agent_id = coder_agent.dev.id
+  url      = "${data.coder_parameter.repo.value}"
+}
 
+module "coder-login" {
+  count    = data.coder_workspace.me.start_count
+  source   = "registry.coder.com/modules/coder-login/coder"
+  agent_id = coder_agent.dev.id
 }
 
 resource "coder_agent" "dev" {
@@ -173,7 +165,7 @@ resource "coder_agent" "dev" {
   }
 
   display_apps {
-    vscode = data.coder_parameter.ide.value == "code"
+    vscode = true
     vscode_insiders = false
     ssh_helper = false
     port_forwarding_helper = true
@@ -183,74 +175,9 @@ resource "coder_agent" "dev" {
   startup_script_behavior = "non-blocking"
   connection_timeout = 300
   startup_script  = <<EOT
-#!/bin/sh
-
-# set -e
-
-if [ "${data.coder_parameter.ide.value}" = "code-server" ]; then
-  # start code-coder
-  # Append "--version x.x.x" to install a specific version of code-server
-
-    curl -fsSL https://code-server.dev/install.sh | sh -s -- --method=standalone --prefix=/tmp/code-server
-
-    # Start code-server in the background.
-    /tmp/code-server/bin/code-server --auth none --port 13337 >/tmp/code-server.log 2>&1 &
-
-fi
-
-# use coder CLI to clone and install dotfiles
-if [ ! -z "${data.coder_parameter.dotfiles_url.value}" ]; then
-  coder dotfiles -y ${data.coder_parameter.dotfiles_url.value}
-fi
-
-# clone repo
-
-echo "folder_name: ${local.folder_name}"
-echo "repo_name: ${local.repo_owner_name}"
-
-if test -z "${data.coder_parameter.repo.value}"
-then
-  echo "No git repo specified, skipping"
-else
-  if [ ! -d "${local.folder_name}" ]
-  then
-    echo "Cloning git repo..."
-    git clone ${data.coder_parameter.repo.value}
-  else
-    echo "directory and repo ${local.folder_name} exists, so skipping clone"
-  fi
-fi
+  #!/bin/sh
 
   EOT
-}
-
-# coder technologies' code-server
-resource "coder_app" "coder-code-server" {
-  count = data.coder_parameter.ide.value == "code-server" ? 1 : 0
-  agent_id = coder_agent.dev.id
-  slug          = "coder"
-  display_name  = "code-server"
-  url      = "http://localhost:13337"
-  icon     = "/icon/code.svg"
-  subdomain = false
-  share     = "owner"
-
-  healthcheck {
-    url       = "http://localhost:13337/healthz"
-    interval  = 5
-    threshold = 15
-  }
-}
-
-# zed ide
-resource "coder_app" "zed" {
-  count = data.coder_parameter.ide.value == "zed" ? 1 : 0
-  agent_id = coder_agent.dev.id
-  slug          = "slug"
-  display_name  = "Zed"
-  external = true
-  url      = "zed://ssh/coder.${data.coder_workspace.me.name}"
-  icon     = "/icon/zed.svg"
 }
 
 resource "docker_container" "workspace" {